refactor for AlibabaResourceReference changes
This change updates the logic around security group IDs, vswitch IDs,
and resource group IDs to match the union changes in the resource
reference API. AlibabaResourceReferences now contain a `Type` variable
which indicates whether it is an ID or a list of Tags.

The behavior has not been changed for security groups, vswitches, and
resource groups. This change is being added to help allow for scenarios
where the ResourceGroupID is not known and must be searched. Now that
the AlibabaResourceReference contains a discriminating union of ID and
Tags, we can properly address scenarios with limited information.