Ensure HOME is set for cache setup
When the HOME directory is unset or is set to /, kubebuilder tries to
create a cache directory but fails. Make sure in these cases that we set
it to a temp folder so that the test can create its cache.

This mimics what we already do in other providers repos.